Refactoring

+++ src/main/java/com/project/springbootwebstore/dto/CartItem.java
@@ -1,7 +1,7 @@
 package com.project.springbootwebstore.dto;
 
 public record CartItem(
-        int id,
-        int quantity
+        long id,
+        long quantity
 ) {
 }
